TagParser.cs was added. Parsing methods from TokenReader.cs were moved to TagParser.cs
